+++ java-daemon/src/main/java/net/cryptic_game/backend/daemon/App.java
@@ -49,6 +49,7 @@ public class App extends AppBootstrap {
                 e.printStackTrace();
             }
         }
+        this.daemonEndpointHandler.postInit();
     }
 
     @Override
